+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.35.0 (Unreleased)
+## 0.35.0 (March 31, 2020)
 FEATURES:
 * **New Resource:** `yandex_kms_secret_ciphertext`